refactor: Follow 도메인을 User 도메인 안에 포함.
