+++ src/main/java/org/ahpuh/surf/user/dto/response/FollowUserResponseDto.java
@@ -1,4 +1,4 @@
-package org.ahpuh.surf.follow.dto;
+package org.ahpuh.surf.user.dto.response;
 
 import lombok.*;
 
@@ -6,7 +6,7 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
 @Builder
-public class FollowUserDto {
+public class FollowUserResponseDto {
 
     private Long userId;
