+++ src/main/java/org/ahpuh/surf/user/service/FollowService.java
@@ -1,14 +1,14 @@
-package org.ahpuh.surf.follow.service;
+package org.ahpuh.surf.user.service;
 
 import lombok.RequiredArgsConstructor;
 import org.ahpuh.surf.common.exception.EntityExceptionHandler;
-import org.ahpuh.surf.follow.domain.Follow;
-import org.ahpuh.surf.follow.domain.FollowConverter;
-import org.ahpuh.surf.follow.domain.FollowRepository;
-import org.ahpuh.surf.follow.dto.FollowUserDto;
-import org.ahpuh.surf.follow.dto.response.FollowResponseDto;
 import org.ahpuh.surf.user.domain.User;
 import org.ahpuh.surf.user.domain.UserRepository;
+import org.ahpuh.surf.user.domain.follow.Follow;
+import org.ahpuh.surf.user.domain.follow.FollowConverter;
+import org.ahpuh.surf.user.domain.follow.FollowRepository;
+import org.ahpuh.surf.user.dto.response.FollowResponseDto;
+import org.ahpuh.surf.user.dto.response.FollowUserResponseDto;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -51,7 +51,7 @@ public class FollowService {
         followRepository.delete(followEntity);
     }
 
-    public List<FollowUserDto> findFollowerList(final Long userId) {
+    public List<FollowUserResponseDto> findFollowerList(final Long userId) {
         final User userEntity = userRepository.findById(userId)
                 .orElseThrow(() -> UserNotFound(userId));
         return followRepository.findByFollowedUser(userEntity)
@@ -61,7 +61,7 @@ public class FollowService {
                 .toList();
     }
 
-    public List<FollowUserDto> findFollowingList(final Long userId) {
+    public List<FollowUserResponseDto> findFollowingList(final Long userId) {
         final User userEntity = userRepository.findById(userId)
                 .orElseThrow(() -> UserNotFound(userId));
         return followRepository.findByUser(userEntity)
