+++ src/test/java/org/ahpuh/surf/integration/follow/controller/FollowControllerTest.java
@@ -1,11 +1,11 @@
 package org.ahpuh.surf.integration.follow.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.ahpuh.surf.follow.domain.Follow;
-import org.ahpuh.surf.follow.domain.FollowRepository;
 import org.ahpuh.surf.user.controller.UserController;
 import org.ahpuh.surf.user.domain.User;
 import org.ahpuh.surf.user.domain.UserRepository;
+import org.ahpuh.surf.user.domain.follow.Follow;
+import org.ahpuh.surf.user.domain.follow.FollowRepository;
 import org.ahpuh.surf.user.dto.request.UserLoginRequestDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
